// src/api/models/PlayerGameCount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PlayerGameCount
 */
export interface PlayerGameCount {
    /**
     * 
     * @type {string}
     * @memberof PlayerGameCount
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof PlayerGameCount
     */
    count: number;
}

/**
 * Check if a given object implements the PlayerGameCount interface.
 */
export function instanceOfPlayerGameCount(value: object): value is PlayerGameCount {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    return true;
}

export function PlayerGameCountFromJSON(json: any): PlayerGameCount {
    return PlayerGameCountFromJSONTyped(json, false);
}

export function PlayerGameCountFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlayerGameCount {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'count': json['count'],
    };
}

export function PlayerGameCountToJSON(json: any): PlayerGameCount {
    return PlayerGameCountToJSONTyped(json, false);
}

export function PlayerGameCountToJSONTyped(value?: PlayerGameCount | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'count': value['count'],
    };
}